Add per item screenshot tests
After trying to debug some screenshot tests it became clear they needed
to be more granular. This is a partial change adding the framework for
that. Screenshots are from version v4.0.0 of the AAI executable. These
tests seem to be robust to reruns.

(note that the multi arena tests will fail with v4.0.0, as they are
included in the subsequent version)

Also:
- Add closing aai if the assert fails in run_behaviour_in_aai
- Add a .gitignore