Update Tests for mlagents upgrade
The camera view in the new version of mlagents is transposed from the
previous one, so this update required a transpose.

There were also non-transposition differences in the test screenshots,
so I have verified they are all not human-visible and updated all the
.pkl files.

The badgoalmulti screenshot test was also broken prior to this change
(the object did not appear at all). It now works, perhaps because of https://github.com/Kinds-of-Intelligence-CFI/animal-ai-unity/commit/058c71de2e752b67db75f21abadc0c3717fa7d55,
so updating the screenshot tests also encompasses this fix.