Update screenshot tests for the github action linux screenshots
Screenshot tests will be probably always vary from platform to platform due to e.g. different shaders being available. Whatever screenshots are generated in CI should be the 'ground truth' for this, as they're the ones we don't want to have to manually check when they fail. Currently this makes the screenshots Linux specific, but a later change will add an environment variable that configures the tests for different platforms